+++ geomars-pytorch/hparams.py
@@ -2,16 +2,16 @@ HASH_BITS = 128
 LAMBDA1 = 0.001
 LAMBDA2 = 1
 LR = 0.0001 #0.0001
-BATCH_SIZE = 8
-EPOCHS = 1
+BATCH_SIZE = 256
+EPOCHS = 30
 DENSENET_NUM_FEATURES = 1024
 ADAM_BETAS = (0.5,0.7) #(0.5,0.7)
 MARGIN = 0.2 #0.2
 INTERCLASSTRIPLETS = False
 KMEANS_CLUSTERS = 30
-MULTIVIEWS=False
-PROJ_DIM = 256
-TEMPERATURE = 0.1
 DENSENET_TYPE = "imagenet" # "domars16k_classifier", "domars16k_triplet"
 DOMAIN_ADAPTION = False
 DA_GROUP_SIZE = DENSENET_NUM_FEATURES // 32
+MULTIVIEWS=False
+PROJ_DIM = 256
+TEMPERATURE = 0.1
